:lipstick: Move test extensions to correct folder
